pyproject.toml
[tool.poetry]
name = "deprep"
version = "0.1.0"
description = ""
authors = ["Pete Fein <pete@snake.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
PyGithub = "^1.47"
environs = "^7.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
